Updated database schema

+++ librarian_football/hooks.py
@@ -1,5 +1,6 @@
 import os, json
 from .menuitems import FootballMenuItem
+from .football import *
 
 def initialize(supervisor):
     supervisor.exts.menuitems.register(FootballMenuItem)
@@ -11,6 +12,11 @@ def post_start(supervisor):
     fsal_client = supervisor.exts.fsal
     (success, dirs, files) = fsal_client.list_dir(config['football.data_dir'])
 
+    if files is None:
+        # no new data, so just return
+        return
+
+    dump_database(db)
     for f in files:
         if f.name == 'leagues_list.json':
             q_delete = db.Delete('leagues')
@@ -19,10 +25,16 @@ def post_start(supervisor):
             with open(path) as json_data:
                 data = json.load(json_data)
                 for league in data:
-                    q_insert = db.Insert('leagues', cols=['id', 'name', 'created'])
+                    q_insert = db.Insert('leagues', cols=['id',
+                                                          'name', 
+                                                          'created',
+                                                          'current_matchday',
+                                                          'number_of_matchdays'])
                     vals = { "id": league['id'],
                              "name": league['caption'],
-                             "created": league['lastUpdated']
+                             "created": league['lastUpdated'],
+                             "current_matchday": league['currentMatchday'],
+                             "number_of_matchdays": league['numberOfMatchdays']
                     }
                     db.execute(q_insert, vals)
             break
